Run the cache conformance suite on Postgres and MySQL, and fix what it caught
The database store was only ever exercised on SQLite. Pointing the same suite
at real servers failed twelve tests on one cause: `expiration` was a 32-bit
integer, and "forever" is written as a timestamp past 2038. Postgres and MySQL
both refuse it as out of range; SQLite stores it happily, which is why a store
that had been passing for months was broken on both servers people deploy to.

The column is a big integer now, in the `cache:table` stub, in the playground's
migration and in the suite's own tables. Nothing else changed — the store's SQL
was right, the schema it was written against was not.

The suite provisions its own tables per run so two runs on one server never
collide, and skips with a note when a server is not there.

The smoke check for the funnel was also asserting the wrong thing: it wanted a
peak of exactly two, which tests how fast a driver hands out a slot rather than
the semaphore. On the file store the first caller can finish before the second
gets in, and that is not a fault. It now asserts the invariant — never more
than the limit, and some callers refused rather than queued.

Recorded in the watch list: `sessions.last_activity` is the same 32-bit shape
and will meet the same wall in 2038.

1599 tests, 533 smoke checks.

+++ playground/database/migrations/2026_08_12_062107_create_cache_table.ts
@@ -13,14 +13,14 @@ export default class extends Migration {
     await schema.create('cache', (table) => {
       table.string('key').primary()
       table.text('value')
-      table.integer('expiration')
+      table.bigInteger('expiration')
       table.index(['expiration'])
     })
 
     await schema.create('cache_locks', (table) => {
       table.string('key').primary()
       table.string('owner')
-      table.integer('expiration')
+      table.bigInteger('expiration')
     })
   }
